backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.13.1
pyjwt==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
pillow==11.2.1
boto3==1.34.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.2.1
